Add /users, /mods, /admins
Oh come on, /users is more of a natural command than /ol, I'm surprised that /users isn't one of the commands for summoning the online users list. Add it as an additional method of calling it up.

Also adding the list of mods and admins, just edit it whenever you make a mod. I'd hope that it doesn't become a hassle, because if it does, YOU'RE MAKING TOO MANY OF THEM.